add some annotation and change  variable name
add some annotation and change  variable name
